--- runp.py
@@ -1,3 +1,0 @@
-#!flask/bin/python
-from app import app
-app.run(host='0.0.0.0')
